backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.7.0
python-dotenv>=1.0.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
